Split scores export precedence from materialization helpers.
Move snapshot types, branch classifier, and unified resolver into export_precedence.py; keep wire shaping and gathering in export_materialization with re-exports for stable imports.
